+++ lib/src/core/services/exceptions/server/server_error_mapper.dart
@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 import 'package:dio/dio.dart';
 
 import 'server_error_exception.dart';
@@ -7,6 +9,8 @@ import 'server_error_exception.dart';
 class ServerErrorMapper {
   /// Для обработки остальных серверных ошибок
   /// необходимо задать новые целочисленные константы, соотвтетсвующие кодам.
+  static const int _badRequest = 400;
+  static const int _unauthorized = 401;
   static const int _notFound = 404;
 
   static Exception fromDioError(DioError error) {
@@ -15,8 +19,19 @@ class ServerErrorMapper {
     /// Для обработки остальных серверных ошибок
     /// нужно написать дополнительные блоки в условном выражении,
     /// предварительно создав классы исключений.
-    if (statusCode == _notFound) {
-      return NotFoundException();
+    switch (statusCode) {
+      case _notFound:
+        return NotFoundException();
+      case _unauthorized:
+        return UnauthorizedException();
+      case _badRequest:
+        if (error.response?.toString() ==
+            jsonEncode({
+              'code': 101,
+            })) {
+          return ImageSoLargeException();
+        }
+        break;
     }
 
     /// Возвращается по-умолчанию для остальных ошибок